FE: add alert when delete task

+++ client/components/task/index.tsx
@@ -11,8 +11,8 @@ import { Clock, Trash2, Eye, Pencil } from "lucide-react";
 import { Button } from "../ui/button";
 
 import moment from "moment";
-import { useDispatch } from "react-redux";
-import { AppDispatch } from "@/redux/store";
+import { useDispatch, useSelector } from "react-redux";
+import { AppDispatch, RootState } from "@/redux/store";
 
 import { DetailTaskModal } from "../modal/detailTaskModal";
 import { DialogTrigger } from "@radix-ui/react-dialog";
@@ -20,6 +20,18 @@ import { deleteTask } from "@/redux/taskSlice";
 import { toast } from "react-toastify";
 
 import { EditTaskModal } from "@/components/modal/editTaskModal.jsx";
+import {
+  AlertDialog,
+  AlertDialogAction,
+  AlertDialogCancel,
+  AlertDialogContent,
+  AlertDialogDescription,
+  AlertDialogFooter,
+  AlertDialogHeader,
+  AlertDialogTitle,
+  AlertDialogTrigger,
+} from "../ui/alert-dialog";
+import { CategoryType } from "@/app/category/[slug]/page";
 
 type InputType = {
   task: TaskType;
@@ -34,6 +46,11 @@ const PriorityBadge = ({ num }: { num: number }) => {
 
 const TaskComponent = ({ task }: InputType) => {
   const dispatch = useDispatch<AppDispatch>();
+
+  const categories = useSelector((state: RootState) => state.categories);
+  const category: CategoryType = categories.data?.filter(
+    (item: CategoryType) => item._id === task.category
+  )[0];
 
   const handleDeleteTask = (id: string) => {
     dispatch(deleteTask(id))
@@ -73,12 +90,35 @@ const TaskComponent = ({ task }: InputType) => {
               </span>
             </DialogTrigger>
           </DetailTaskModal>
-          <Button
-            onClick={() => handleDeleteTask(task._id)}
-            className="bg-background hover:bg-secondary text-foreground"
-          >
-            <Trash2 />
-          </Button>
+          <AlertDialog>
+            <AlertDialogTrigger asChild>
+              <Button className="bg-background hover:bg-secondary text-foreground">
+                <Trash2 />
+              </Button>
+            </AlertDialogTrigger>
+            <AlertDialogContent>
+              <AlertDialogHeader>
+                {" "}
+                <AlertDialogTitle>
+                  ⚠️ Are you absolutely sure?
+                </AlertDialogTitle>{" "}
+                <AlertDialogDescription>
+                  {" "}
+                  This action cannot be undone. This will permanently delete
+                  your <b>category</b> and remove all its <b>tasks</b>.
+                </AlertDialogDescription>{" "}
+              </AlertDialogHeader>
+              <AlertDialogFooter>
+                <AlertDialogCancel>Cancel</AlertDialogCancel>
+                <AlertDialogAction
+                  className="bg-red-400"
+                  onClick={() => handleDeleteTask(task._id)}
+                >
+                  Delete
+                </AlertDialogAction>
+              </AlertDialogFooter>
+            </AlertDialogContent>
+          </AlertDialog>
         </div>
       </CardHeader>
       <CardContent>
@@ -107,9 +147,22 @@ export const MiniTaskComponent = ({ task }: InputType) => {
         </Button>
       </CardHeader>
       <CardFooter className="flex justify-between">
-        <div className="text-priority-high flex flex-row items-center space-x-2">
-          <Clock />
-          <div>
+        <div className="flex gap-3">
+          <div className="flex items-center">
+            <div
+              className={`w-3 h-3 rounded-full`}
+              style={{ backgroundColor: category.color }}
+            ></div>
+            <span
+              className={
+                "font-medium w-fit ml-3 text-sm overflow-hidden transition-all"
+              }
+            >
+              {category.name}
+            </span>
+          </div>
+          <div className="text-priority-high flex flex-row items-center space-x-2 gap-2">
+            <Clock />
             {moment(new Date(task.end_day)).format("DD/MM/YYYY hh:mm:ss")}
           </div>
         </div>
